Convert LogRecord to use RecordLength

+++ libs/storage/Tsavorite/cs/src/core/Index/Recovery/IndexRecovery.cs
@@ -169,7 +169,7 @@ namespace Tsavorite.core
                             b->bucket_entries[bucket_entry] = 0;
                     }
                     // Reset any ephemeral bucket level locks
-                    b->bucket_entries[Constants.kOverflowBucketIndex] &= LogAddress.kAddressBitMask;
+                    b->bucket_entries[Constants.kOverflowBucketIndex] &= (long)LogAddress.kAddressBitMask;
                     if (b->bucket_entries[Constants.kOverflowBucketIndex] == 0) break;
                     b = (HashBucket*)overflowBucketsAllocator.GetPhysicalAddress(b->bucket_entries[Constants.kOverflowBucketIndex]);
                 }
